do not use full event monitor
